# swagger.yml
basePath: /user-service/api/v1
swagger: '2.0'
info:
  version: 1.0.0
  title: User Service
  description: Managing Users and Authentication
  termsOfService: terms
  contact:
    name: David Bohl
    url: https://github.com/DerDavidBohl
    email: david@bohl24.de
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
paths:
  /authorize/code:
    get:
      parameters:
        - $ref: '#/parameters/redirectUriInQuery'
      responses:
        '301':
          description: Redirect to RiderectUri
  /authorize/token:
    get:
      parameters:
        - $ref: '#/parameters/tokenCodeInQuery'
      responses:
        '200':
          description: Token in Body and Headers (UserService-Application Role Needed)
  /login:
    post:
      security: []
      description: Authenticates a User
      parameters:
        - $ref: '#/parameters/loginInBody'
      responses:
        '200':
          description: OK
          headers:
            authorization:
              type: string
  /reset:
    post:
      security: []
      description: Resets a password for a email
      parameters:
        - $ref: '#/parameters/emailInQuery'
        - $ref: '#/parameters/resetTokenInQuery'
        - $ref: '#/parameters/newPasswordInBody'
      responses:
        '204':
          description: OK
  /reset/request:
    post:
      security: []
      description: Requests a password reset for a email
      parameters:
        - $ref: '#/parameters/emailInBody'
      responses:
        '204':
          description: OK
  /users:
    get:
      description: Gets all Users
      parameters:
        - $ref: '#/parameters/roleInQuery'
      responses:
        '200':
          description: OK - With User
    post:
      description: Creates a new User
      parameters:
        - $ref: '#/parameters/userInBody'
      responses:
        '201':
          description: Created
  /users/verify:
    post:
      security: []
      parameters:
        - $ref: '#/parameters/emailInQuery'
        - $ref: '#/parameters/verifyTokenInQuery'
      description: Verifies a User
      responses:
        '200':
          description: OK
  /users/current:
    get:
      description: gets the current User
      responses:
        '200':
          description: OK
  /users/current/password:
    put:
      description: Sets the password for the current User
      parameters:
        - $ref: '#/parameters/newPasswordInBody'
      responses:
        '204':
          description: Password set
  /users/current/roles:
    get:
      description: Gets current Users Roles
      responses:
        '200':
          description: OK
          schema:
            type: array
            items: 
              type: string
  /users/{userId}:
    parameters:
      - $ref: '#/parameters/userIdInPath'
    get:
      description: Gets a specific User
      responses:
        '200':
          description: OK
    delete:
      description: Deletes a User
      responses:
        '204':
          description: Deleted
  # /users/{userId}/data:
  #   parameters:
  #     - $ref: '#/parameters/userIdInPath'
  /users/{userId}/roles:
    parameters:
      - $ref: '#/parameters/userIdInPath'
    get:
      description: Gets a specific Users Roles
      responses:
        '200':
          description: OK
          schema:
            type: array
            items: 
              type: string
    post:
      description: Adds a Role to a User
      parameters:
        - $ref: '#/parameters/rolesInBody'
      responses:
        '201':
          description: OK - Role added         
  /users/{userId}/roles/{role}:
    parameters:
      - $ref: '#/parameters/userIdInPath'
      - $ref: '#/parameters/roleInPath'
    delete:
      description: Removes a Role from a User
      responses:
        '204':
          description: OK
parameters:
  tokenCodeInQuery:
    name: code
    in: query
    type: string
    description: code for Token authorization
  redirectUriInQuery:
    name: redirect
    in: query
    type: string
    description: Redirect Uri in Query
    default: ./dummy
  resetTokenInQuery:
    name: token
    in: query
    type: string
    description: reset token in Query
    required: true
  emailInQuery:
    name: email
    in: query
    type: string
    description: email address in Query
    required: true
  newPasswordInBody:
    name: newPasswordInBody
    in: body
    required: true
    description: new password in body
    schema:
      $ref: '#/definitions/NewPassword'
  emailInBody:
    name: emailInBody
    in: body
    description: email-address in body
    schema:
      $ref: '#/definitions/Email'
  userIdInPath:
    name: userId
    in: path
    required: true
    type: string
    description: Id of an User
  roleInPath:
    name: role
    in: path
    required: true
    type: string
    description: Id of an User
  roleInQuery:
    name: role
    in: query
    type: string
    description: Role Filter
  rolesInBody:
    name: rolesInBody
    in: body
    description: Role in Body
    schema:
      $ref: '#/definitions/Roles'
  userInBody:
    name: userInBody
    in: body
    description: User Object in Body
    schema:
      $ref: '#/definitions/User'
  loginInBody:
    name: loginInBody
    in: body
    description: Login in Body
    schema:
      $ref: '#/definitions/Login'
  verifyTokenInQuery:
    name: token
    required: true
    in: query
    type: string
definitions:
  User:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
  Email:
    type: object
    properties:
      email:
        type: string
  Roles:
    type: array
    items:
      type: string
  Login:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
  NewPassword:
    type: object
    properties:
      password:
        type: string
securityDefinitions:
  Bearer:
    type: apiKey
    name: authorization
    in: header
security:
  - Bearer: []